# API-EXERCISE/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Golang API.",
    "version": "0.0.1-alpha"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "paths": {
...
    "/titles/{titleId}/credits": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "Credits"
        ],
        "summary": "List credits by titleId.",
        "operationId": "RequestGetCreditTitle",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TitleId",
            "name": "titleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListCredit"
          },
          "404": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "Credits"
        ],
        "summary": "Create a credit.",
        "operationId": "RequestCreateCredit",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TitleId",
            "name": "titleId",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "character": {
                  "type": "string",
                  "x-go-name": "Character"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "person_id": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "PersonID"
                },
                "role": {
                  "type": "string",
                  "x-go-name": "Role"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/ResponseCredit"
          },
          "400": {
            "$ref": "#/responses/GenericResFailBadRequest"
          },
          "404": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/titles/{titleId}/credits/{creditId}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "Credits"
        ],
        "summary": "Get a credit.",
        "operationId": "RequestGetCredit",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TitleId",
            "name": "titleId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "CreditId",
            "name": "creditId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseCredit"
          },
          "404": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "Credits"
        ],
        "summary": "Delete a credit.",
        "operationId": "RequestDeleteCredit",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TitleId",
            "name": "titleId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "CreditId",
            "name": "creditId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseDeleteCredit"
          },
          "404": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
  },
  "responses": {
    "GenericResError": {
      "description": "Unexpected error occurred",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "enum": [
              "error"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
    "GenericResFailBadRequest": {
      "description": "Fail due to user invalid input",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": [
              "fail"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "GenericResFailNotFound": {
      "description": "Fail due to resource not exists",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": [
              "fail"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseCredit": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "character": {
                "type": "string",
                "x-go-name": "Character"
              },
              "created_at": {
                "type": "string",
                "x-go-name": "CreatedAt"
              },
              "id": {
                "type": "string",
                "x-go-name": "Id"
              },
              "name": {
                "type": "string",
                "x-go-name": "Name"
              },
              "person_id": {
                "type": "integer",
                "format": "int64",
                "x-go-name": "PersonID"
              },
              "role": {
                "type": "string",
                "x-go-name": "Role"
              },
              "title_id": {
                "type": "string",
                "x-go-name": "TitleID"
              },
              "updated_at": {
                "type": "string",
                "x-go-name": "UpdatedAt"
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
    "ResponseDeleteCredit": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseListCredit": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "character": {
                  "type": "string",
                  "x-go-name": "Character"
                },
                "created_at": {
                  "type": "string",
                  "x-go-name": "CreatedAt"
                },
                "id": {
                  "type": "string",
                  "x-go-name": "Id"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "person_id": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "PersonID"
                },
                "role": {
                  "type": "string",
                  "x-go-name": "Role"
                },
                "title_id": {
                  "type": "string",
                  "x-go-name": "TitleID"
                },
                "updated_at": {
                  "type": "string",
                  "x-go-name": "UpdatedAt"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
  }
}